+++ src/fides/api/util/tcf/tc_mobile_data.py
@@ -5,7 +5,8 @@ from iab_tcf import ConsentV2, decode_v2  # type: ignore[import]
 
 from fides.api.common_exceptions import DecodeFidesStringError
 from fides.api.schemas.tcf import TCMobileData
-from fides.api.util.tcf.ac_string import build_ac_string, split_fides_string
+from fides.api.util.tcf.ac_string import build_ac_string
+from fides.api.util.tcf.fides_string import split_fides_string
 from fides.api.util.tcf.tc_model import TCModel
 from fides.api.util.tcf.tc_string import (
     PURPOSE_CONSENTS_BITS,
