+++ src/fides/config/consent_settings.py
@@ -1,4 +1,6 @@
-from pydantic import Field
+from typing import Any, Dict
+
+from pydantic import Field, root_validator
 
 from .fides_settings import FidesSettings
 
@@ -15,3 +17,14 @@ class ConsentSettings(FidesSettings):
 
     class Config:
         env_prefix = "FIDES__CONSENT__"
+
+    @root_validator
+    def validate_fields(cls, values: Dict[str, Any]) -> Dict[str, Any]:
+        """AC mode only works if TCF mode is also enabled"""
+        tcf_mode = values.get("tcf_enabled")
+        ac_mode = values.get("ac_enabled")
+
+        if ac_mode and not tcf_mode:
+            raise ValueError("TCF must be enabled in order for AC features to work.")
+
+        return values
